+++ sql-logic-service/src/main/java/com/sql/logic/engine/domain/agent/node/SqlGenerationNode.java
@@ -65,11 +65,11 @@ public class SqlGenerationNode implements NodeAction {
         String evidence = state.value(SqlAgentSpec.StateKey.EVIDENCE, "");
         String existingTableRelation = state.value(SqlAgentSpec.StateKey.TABLE_RELATION, "");
 
-        Object connectionIdObj = state.value(SqlAgentSpec.StateKey.CONNECTION_ID, null);
+        Object connectionIdObj = state.value(SqlAgentSpec.StateKey.CONNECTION_ID, (Long) null);
         Long connectionId = AgentStateUtil.toLong(connectionIdObj);
-        Object llmConfigIdObj = state.value(SqlAgentSpec.StateKey.LLM_CONFIG_ID, null);
+        Object llmConfigIdObj = state.value(SqlAgentSpec.StateKey.LLM_CONFIG_ID, (Long) null);
         Long llmConfigId = AgentStateUtil.toLong(llmConfigIdObj);
-        Object userIdObj = state.value(SqlAgentSpec.StateKey.USER_ID, null);
+        Object userIdObj = state.value(SqlAgentSpec.StateKey.USER_ID, (Long) null);
         Long userId = AgentStateUtil.toLong(userIdObj);
 
         // If no rewritten query, fall back to original input
